api/requirements.txt
sqlalchemy
pydantic[email]
fastapi
passlib[bcrypt]
uvicorn
python-jose